@@ -4,21 +4,22 @@
   "description": "",
   "main": "index.js",
   "dependencies": {
-    "rxjs": "^5.0.3"
+    "rxjs": "^6.4.0"
   },
   "devDependencies": {
-    "babel-core": "^6.22.1",
-    "babel-loader": "^6.2.10",
-    "http-server": "^0.9.0",
-    "json-server": "^0.9.6",
-    "npm-run-all": "^4.0.2",
-    "webpack": "^2.2.1"
+    "babel-core": "^6.26.3",
+    "babel-loader": "^8.0.5",
+    "http-server": "^0.11.1",
+    "json-server": "^0.14.2",
+    "npm-run-all": "^4.1.5",
+    "webpack": "^4.29.6",
+    "webpack-cli": "^3.3.0"
   },
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
     "start": "npm-run-all --parallel webpack http-server api",
     "webpack": "webpack -w",
-    "http-server": "http-server",
+    "http-server": "http-server -c-1",
     "api": "json-server ./4_from_promise/db/db.json -d 2000"
   },
   "author": "Lehel Vass",
