@@ -1,12 +1,11 @@
-'use strict';
-import Rx from 'rxjs/Rx';
+import { from } from 'rxjs';
 
-// Create Observable from value(s)
-// Passing string vs array
-const fromString$ = Rx.Observable.from('hello');
-const fromArray$ = Rx.Observable.from(['hello']);
-const fromObject$ = Rx.Observable.from([{key: 'hello'}]);
+// // Create Observable from value(s)
+// // Passing string vs array
+const fromString$ = from('hello');
+const fromArray$ = from(['hello']);
+const fromObject$ = from([{ key: 'hello' }]);
 
-fromString$.subscribe((res) => console.log(res));
-fromArray$.subscribe((res) => console.log(res));
-fromObject$.subscribe((res) => console.log(res));
+fromString$.subscribe(res => console.log(res));
+fromArray$.subscribe(res => console.log(res));
+fromObject$.subscribe(res => console.log(res));
