@@ -1,10 +1,9 @@
-'use strict';
-const Rx = require('rxjs/Rx');
+import { Subject, ReplaySubject } from 'rxjs';
 
 class RxStore {
   constructor() {
-    // this.store = new Rx.Subject();
-    this.store = new Rx.ReplaySubject();
+    // this.store = new Subject();
+    this.store = new ReplaySubject();
   }
 
   dispatch(message) {
@@ -18,9 +17,9 @@ class RxStore {
 
 const myStore = new RxStore();
 
-// difference between Subject and ReplaySubject - demonstarte with changing dispatch and subscribe order
+// // difference between Subject and ReplaySubject - demonstarte with changing dispatch and subscribe order
 myStore.dispatch('Hello');
-myStore.dispatch('World');  // It's an accumulator
+myStore.dispatch('World'); // It's an accumulator
 
 myStore.subscribe(v => console.log(v));
 myStore.subscribe(v => console.log('Second subscriber ' + v));
